# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pricing-lab",
    "description": "Causal-first dynamic pricing lab \u2014 switchback designs, ATE estimators.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/simulate": {
      "post": {
        "tags": [
          "simulation"
        ],
        "summary": "Simulate",
        "operationId": "simulate_v1_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/estimate/naive": {
      "post": {
        "tags": [
          "estimation"
        ],
        "summary": "Estimate Naive",
        "operationId": "estimate_naive_v1_estimate_naive_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstimateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Estimate Naive V1 Estimate Naive Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/estimate/switchback": {
      "post": {
        "tags": [
          "estimation"
        ],
        "summary": "Estimate Switchback",
        "operationId": "estimate_switchback_v1_estimate_switchback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstimateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Estimate Switchback V1 Estimate Switchback Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/compare": {
      "post": {
        "tags": [
          "estimation"
        ],
        "summary": "Compare",
        "operationId": "compare_v1_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstimateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CompareResponse": {
        "properties": {
          "true_ate": {
            "type": "number",
            "title": "True Ate"
          },
          "estimators": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Estimators"
          },
          "cfg": {
            "additionalProperties": true,
            "type": "object",
            "title": "Cfg"
          }
        },
        "type": "object",
        "required": [
          "true_ate",
          "estimators",
          "cfg"
        ],
        "title": "CompareResponse"
      },
      "ConfigOverrides": {
        "properties": {
          "n_zones": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 64.0,
                "minimum": 2.0
              },
              {
                "type": "null"
              }
            ],
            "title": "N Zones"
          },
          "n_time_buckets": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 720.0,
                "minimum": 24.0
              },
              {
                "type": "null"
              }
            ],
            "title": "N Time Buckets"
          },
          "spillover_strength": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Spillover Strength"
          },
          "price_control": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Control"
          },
          "price_treatment": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Treatment"
          },
          "switchback_block_hours": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 72.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Switchback Block Hours"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          }
        },
        "type": "object",
        "title": "ConfigOverrides",
        "description": "Subset of MarketplaceConfig knobs exposed via the API.\n\nAnything omitted falls back to the dataclass default."
      },
      "EstimateRequest": {
        "properties": {
          "overrides": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfigOverrides"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "EstimateRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "status",
          "version"
        ],
        "title": "HealthResponse"
      },
      "SimulateRequest": {
        "properties": {
          "design": {
            "type": "string",
            "title": "Design",
            "description": "`ab_random` or `switchback`",
            "default": "ab_random"
          },
          "overrides": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfigOverrides"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "SimulateRequest"
      },
      "SimulateResponse": {
        "properties": {
          "n_rows": {
            "type": "integer",
            "title": "N Rows"
          },
          "true_ate_revenue": {
            "type": "number",
            "title": "True Ate Revenue"
          },
          "cfg": {
            "additionalProperties": true,
            "type": "object",
            "title": "Cfg"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "n_rows",
          "true_ate_revenue",
          "cfg",
          "summary"
        ],
        "title": "SimulateResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}